requirements.txt
-r requirements-common.txt
gunicorn==18.0
psycopg2==2.5.1
django-postgrespool==0.2.4
pyelasticsearch==0.6
django-redis-cache==0.10.0
raven==1.9.4
django-storages>=1.1.8
boto==2.15.0
gevent==0.13.8
hiredis>=0.1.1
dj-database-url==0.2.2
dj-static==0.0.5
django-recaptcha==0.0.6
django-secure==1.0
static==0.4
